# docs/openapi.yaml
openapi: 3.1.1
info:
  title: API Lanchonete
  version: 1.0.0
  description: >
    API REST para gerenciamento de pedidos,
    produtos, estoque, usuários, pagamentos
    e unidades da lanchonete.
servers:
  - url: https://lanchonete-api.onrender.com
    description: Servidor de produção

  - url: http://127.0.0.1:5000
    description: Servidor local
...
paths:
...
  /login/pedidos/{pedidoId}:
    post:
      tags:
      - Pagamentos
      summary: Rota para pagamento
      description: Permite que o cliente realize o pagamento de um pedido.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - metodo
              properties:
                metodo:
                  type: string
                  enum:
                  - DINHEIRO
                  - DEBITO
                  - CREDITO
                  - PIX
                  - VALE
      responses:
        "400":
          description: Requisição inválida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "JSON_INVALIDO"
                message: "O corpo da requisição deve ser um JSON válido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/login/pedidos/2"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "422":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "METODO_INVALIDO"
                message: "O método de pagamento informado é inválido."
                details: [{field: "metodopagamento", issue: "Valores válidos: PIX, DEBITO, CREDITO, VALE, DINHEIRO."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/login/pedidos/2"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/login/pedidos/2"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        '200':
          description: Pedido pago com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "O pedido foi pago com sucesso."
                data:
                  usuarioId: 2
                  pedidoId: 2
                  metodoPagamento: PIX
                  aprovado: true
                  codigo: "d9e3bb89-2"
        '401':
          description: Token JWT ausente, expirado ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Usuário autenticado sem permissão para acessar este recurso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      operationId: pagamento
      parameters:
      - name: pedidoId
        in: path
        required: true
        description: ID do pedido que será pago.
        schema:
          type: integer
          minimum: 1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          example: "USUARIO_NAO_ENCONTRADO"
        message:
          type: string
          example: "O usuário informado não foi encontrado."
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "usuario"
              issue:
                type: string
                example: "O usuário informado não existe."
        timestamp:
          type: string
          format: date-time
          example: "2026-08-07T08:30:00-03:00"
        path:
          type: string
          example: "/login"
        requestId:
          type: string
          format: uuid
          example: "baaf6e1b-5217-41af-9877-2298562e65e3"
...
    RespostaSucesso:
      type: object
      properties:
        message:
          type: string
          example: Operação realizada com sucesso.
        data: {}
...